mtd: msm_qpic_nand: An incorrect page memset during ECC errors
After every nand read page opertion, we check for ECC errors on the
page that read. If any ECC errors are detected, then we check if the
page read was an erased page, if so we will capture this erased page
in an array and later we will fill the same erased page with 0xff.

So, while filling the array we should add the actual current page that
had bitflips into array rather adding the effective page count.

Also while memsetting the erased page, it should always set the
actual erased page to 0xff rather setting the start page to 0xff.

Change-Id: I20dd109e96860a232100aa6bc4de2c665e85247d
Signed-off-by: Pradeep P V K <ppvk@codeaurora.org>